refactor: 테스트 클래스 패키지 product 코드 패키지와 맞추기
