+++ src/test/kotlin/blackjack/domain/CardTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 import blackjack.domain.Card
 import blackjack.domain.Rank
