+++ src/test/kotlin/blackjack/domain/PlayerTest.kt
@@ -1,11 +1,5 @@
-package blackjack
+package blackjack.domain
 
-import blackjack.domain.Card
-import blackjack.domain.Cards
-import blackjack.domain.Dealer
-import blackjack.domain.Player
-import blackjack.domain.Rank
-import blackjack.domain.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
